AI-OCR/requirements.txt
python-dotenv>=1.1.0
anthropic>=0.17.0
pillow>=10.0.0